requirements.txt
flask
biopython
googletrans==4.0.0-rc1
pandas
